Added MMC3. Can read all PRG and CHR banks but there are likely still problems with the IRQ functionality. Tested with Kirby and game is massively buggy. Changed all return types of PPU read functions in Mappers to 32 bit ints to accomodate game sizes of up to 256Kb.

+++ src/CxROM.cpp
@@ -11,7 +11,7 @@ inline uint32_t CxROM::CPU_READ(uint16_t ADDR) {
 }
 
 
-inline uint16_t CxROM::PPU_READ(uint16_t ADDR, bool NT) {
+inline uint32_t CxROM::PPU_READ(uint16_t ADDR, bool NT) {
 
     return (NT) ? SelectNameTable(ADDR, NT_MIRROR) : CHR_BANK * 2 * CHR_BANK_SIZE + ADDR;
 
@@ -19,7 +19,7 @@ inline uint16_t CxROM::PPU_READ(uint16_t ADDR, bool NT) {
 
 
 inline void CxROM::CPU_WRITE(uint16_t ADDR, uint8_t VAL) {
-    
+    //CNROM only uses the bottom 2 bits capping CHR at 32KB, games using this board with >4 CHR banks should use bottom 4 or more bits
     CHR_BANK = VAL & 0x03;
 }
